Many more clang-tidy fixes.

+++ applications/witness/WitnessThread.cpp
@@ -69,7 +69,7 @@ void WitnessThread::on_connection_lost(pubsub_itc_fw::ConnectionID id, const std
 
 void WitnessThread::on_framework_pdu_message(const pubsub_itc_fw::EventMessage& message) {
     const pubsub_itc_fw::ConnectionID& conn_id = message.connection_id();
-    const auto pdu_id = static_cast<int16_t>(message.pdu_id());
+    const auto pdu_id = message.pdu_id();
 
     if (pdu_id == pdu_arbiter_heartbeat) {
         handle_arbiter_heartbeat(conn_id, message);
